add AsyncChunkDBManager
